Updated almost all methods, added documentation

+++ service/database/deleteLike.go
@@ -1,38 +1,20 @@
 package database
 
 import (
-	"errors"
-	"fmt"
-
 	"github.com/luigidannibale/Wasa/service/utils"
 )
 
+/*
+Errors that can be returned: (NotFound, InternalServerError)
+*/
 func (db *appdbimpl) DeleteLike(like utils.Like) (string, error) {
-	user, s, e := db.GetUser(like.UserID)
-	if e != nil {
-		if e.Error() == "NotFound" {
-			return "Couldn't find the user", errors.New("UserNotFound")
-		}
-		if e.Error() == "InternalServerError" {
-			return s, errors.New("InternalServerError")
-		}
-	}
-	_, s, e = db.GetPhoto(like.PhotoID)
-	if e != nil {
-		if e.Error() == "NotFound" {
-			return "Couldn't find the photo", errors.New("PhotoNotFound")
-		}
-		if e.Error() == "InternalServerError" {
-			return s, errors.New("InternalServerError")
-		}
-	}
-	res, err := db.c.Exec(`DELETE FROM Likes 
-				WHERE UserID = ? AND PhotoID = ?`, like.UserID, like.PhotoID)
+
+	res, err := db.c.Exec(`DELETE FROM Likes WHERE UserID = ? AND PhotoID = ?`, like.UserID, like.PhotoID)
 	if x, y := res.RowsAffected(); x == 0 && y == nil {
-		return fmt.Sprintf("There is no like of %s to the photo %d", user.Username, like.PhotoID), errors.New("BannedNotFound")
+		return "Couldn't find the like", NotFound
 	}
 	if err != nil {
-		return "An error occurred on the server" + err.Error(), errors.New("InternalServerError")
+		return err.Error(), InternalServerError
 	}
-	return fmt.Sprintf("User %s has unliked %d", user.Username, like.PhotoID), nil
+	return "Like deleted successfully", nil
 }
